t/lib-credential: use test_i18ncmp to check stderr
The credential tests have a "check" function which feeds some input to
git-credential and checks the stdout and stderr. We look for exact
matches in the output. For stdout, this makes sense; the output is
the credential protocol. But for stderr, we may be showing various
diagnostic messages, or the prompts fed to the askpass program, which
could be translated. Let's mark them as such.

(cherry picked from commit 17f1c0b8c7e447aa62f85dc355bb48133d2812f2)
Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>

Gbp-Pq: Name 0041-t-lib-credential-use-test_i18ncmp-to-check-stderr.diff