t3432: test `--merge' with `rebase.abbreviateCommands = true', too
commit de9f1d3ef45ec885339d04f9e34293eb2de8605d upstream.

When fast forwarding, `git --merge' should act the same whether
`rebase.abbreviateCommands' is set or not, but so far it was not the
case.  This duplicates the tests ensuring that `--merge' works when fast
forwarding to check if it also works with abbreviated commands.

Signed-off-by: Alban Gruin <alban.gruin@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>
Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>

Gbp-Pq: Name 0003-t3432-test-merge-with-rebase.abbreviateCommands-true-.diff